# openapi-v1.yaml
...
openapi: 3.1.0
info:
  title: Nethical Governance API v1
  description: |
    Enhanced REST API for AI safety and ethics governance with full backend management.
    
    ## New Features in v1
    - **RBAC**: Role-based access control with JWT authentication
      - `admin`: Full access to all operations
      - `auditor`: Read-only access to logs and audit data
      - `operator`: Can evaluate risk, but cannot modify configuration
    - **Agent Management**: CRUD operations for AI agent configuration
    - **Policy Management**: CRUD operations for governance policies
    - **Audit Logs**: Read-only access with Merkle tree verification
    - **Real-time Threats**: WebSocket and SSE for live threat notifications
    
    ## Authentication
    All v1 endpoints (except /auth/login) require authentication via Bearer token.
    
    ### Login Example:
    ```bash
    curl -X POST http://localhost:8000/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"username": "admin", "password": "admin123"}'
    ```
    
    ### Using Token:
    ```bash
    curl http://localhost:8000/api/v1/agents \
      -H "Authorization: Bearer YOUR_TOKEN_HERE"
    ```
    
  version: 1.0.0
  contact:
    name: Nethical Team
    url: https://github.com/V1B3hR/nethical
    email: support@nethical.dev
  license:
    name: MIT
    url: https://github.com/V1B3hR/nethical/blob/main/LICENSE
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
...
paths:
...
  /agents:
    get:
      tags:
        - Agent Management
      summary: List all agents
      description: |
        List all registered AI agents with pagination and filtering.
        
        **Required role:** Any authenticated user
      operationId: listAgents
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [active, suspended, terminated, quarantine]
        - name: agent_type
          in: query
          description: Filter by agent type
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    post:
      tags:
        - Agent Management
      summary: Create new agent
      description: |
        Create a new AI agent configuration.
        
        **Required role:** admin
      operationId: createAgent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
            examples:
              gpt4_agent:
                summary: GPT-4 Agent
                value:
                  agent_id: agent-gpt4-001
                  name: GPT-4 Assistant
                  agent_type: llm
                  description: GPT-4 powered AI assistant
                  trust_level: 0.9
                  status: active
                  configuration:
                    model: gpt-4
                    temperature: 0.7
                    max_tokens: 2000
                  metadata:
                    department: engineering
                    environment: production
      responses:
        '201':
          description: Agent created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Agent with this ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/{agent_id}:
    get:
      tags:
        - Agent Management
      summary: Get agent details
      description: |
        Retrieve details for a specific agent.
        
        **Required role:** Any authenticated user
      operationId: getAgent
      security:
        - BearerAuth: []
      parameters:
        - name: agent_id
          in: path
          required: true
          description: Agent identifier
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Agent not found
    
    patch:
      tags:
        - Agent Management
      summary: Update agent configuration
      description: |
        Update an existing agent's configuration.
        
        **Required role:** admin
      operationId: updateAgent
      security:
        - BearerAuth: []
      parameters:
        - name: agent_id
          in: path
          required: true
          description: Agent identifier
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
      responses:
        '200':
          description: Agent updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Agent not found
    
    delete:
      tags:
        - Agent Management
      summary: Delete agent
      description: |
        Delete an AI agent configuration.
        
        **Required role:** admin
      operationId: deleteAgent
      security:
        - BearerAuth: []
      parameters:
        - name: agent_id
          in: path
          required: true
          description: Agent identifier
          schema:
            type: string
      responses:
        '204':
          description: Agent deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Agent not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    AgentCreate:
      type: object
      required:
        - agent_id
        - name
      properties:
        agent_id:
          type: string
        name:
          type: string
        agent_type:
          type: string
          default: general
        description:
          type: string
        trust_level:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.5
        status:
          type: string
          enum: [active, suspended, terminated, quarantine]
          default: active
        configuration:
          type: object
        metadata:
          type: object
        dock_status:
          type: string
          default: undocked
        visibility:
          type: boolean
          default: false
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        trust_level:
          type: number
        status:
          type: string
        configuration:
          type: object
        metadata:
          type: object
        dock_status:
          type: string
        visibility:
          type: boolean
    AgentResponse:
      type: object
      properties:
        id:
          type: integer
        agent_id:
          type: string
        name:
          type: string
        agent_type:
          type: string
        description:
          type: string
        trust_level:
          type: number
        status:
          type: string
        configuration:
          type: object
        metadata:
          type: object
        dock_status:
          type: string
        visibility:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/AgentResponse'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        pages:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'